requirements.txt
streamlit>=1.32.0
duckdb>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
kaggle>=1.6.0
s3fs>=2024.2.0
boto3>=1.34.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0
